fjerner dokumentClient
