+++ src/main/resources/application.yml
@@ -67,3 +67,4 @@ management:
 
 FAMILIE_OPPDRAG_API_URL: https://familie-oppdrag.prod-fss-pub.nais.io
 ENSLIG_FORSORGER_VEDTAK_TOPIC: teamfamilie.ensligforsorger-vedtak.v1
+FAMILIE_INTEGRASJONER_API_URL: https://familie-integrasjoner.prod-fss-pub.nais.io
